requirements.txt
# Python version: 3.7.9

requests==2.27.1
pyyaml==6.0